move data checks to separate folder
